refactor naming and consts
