+++ src/theme/components/progress.ts
@@ -1,6 +1,6 @@
 import { palette } from 'theme/foundations/colors';
 
-export const progress = {
+const colors = {
   track: palette.blue[500],
   total: palette.neutral.white,
 };
